backend/requirements.txt
weaviate-client==4.8.1
fastapi
fastapi-cli
httpx
uvicorn
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langchainhub
langgraph
langsmith
matplotlib
numpy
openai
pandas
python-socketio
seaborn
tiktoken
websockets
pydantic
pydantic-settings
tavily-python
dependency-injector
# pygraphviz
python-multipart
pytest
